Handle events
As main_quit takes no arguments and the callback registered
for stage destroy event through stage.connect, will recieve the
actor ie stage itself as a argument to it, so a inbetween callback
handle_destroy has been added.

Also added a mouse click handler, which quits the app, if the
actor on which mouse was clicked is the stage.