Some cleanup
Explicitly specify the version of the clutter library being imported,
as suggested by gi warning shown during import of Clutter. This helps
avoid that warning, as well as control the version of the library
that will be used by the program.

Replaced stage.show_all with stage.show.

It would appear that the top level actor (i.e stage in this
case) requires its show function to be called.

	Without this the stage or main app window doesnt appear.

While any children dont require to call show (nor one requires
to call show_all of the top level parent).