Fix a oversight and updated quit key command
As the modifier_state can have multiple flags in it set, so when
checking for its contents one requires to look for

* all flags that could be set simultaneously if checking directly

OR

* or filter the subset that needs to be checked.

The MOD2 mask seems to be set for keys always, so I need to either
add it to the mix of flags being checked for or else I require to
filter|mask it out. This is handled now.

Also I have modified quit key shortcut to Cntrl+Q. Inturn to achieve
the same I need to check for

* Capital Q as well as

* either all of or a subset of the CONTROL, SHIFT and MOD2 Masks in
  the modifier_state.