BoxLayout fixes size in None Orientation axis
If a Actor is using BoxLayout to control the layout of its children,
then the size of the children along the None Orientation axis is
fixed by the Actor's size.

i.e if a Actor is using box layout to arrange its children horizontally
in a row, then the vertical (y-axis) size of children will be same as
the Actor's vertical size. While children can control their own size
along x-axis.

So also, if a Actor is using box layout to arrange its children vertically
into a column, then the horiz (x-axis) size of children will be same as
the Actor's horizontal size. While children can control their own size
along y-axis.