Add a btn press event handler for vertical listbox
NOTE:

Rather the event object passed to btn event handler in python
by default doesnt map to Clutter.Event, but rather it seems to map
to the ButtonEvent structure.

Inturn has this ButtonEvent is not a child of Event, so the functions
in Event cant be called on ButtonEvent directly. Later need to look
into python and gobjects and gi and Clutter/AnyOtherGIInferedObject
relationships and implications.

Some of this could also be due to GObjects being a object based
system written in C and not a native Object oriented system as provided
by C++ or Java, with implicit class heirarchy handling. Havent dabled
at glib, gobject, c++ etc at the lower level for ages now, need to
look at this later.