listbox scrolling - ScrollActor
Use ScrollActor instead of Actor, It clips the view to the size
assigned to it. Which also means that any scaling applied to
its children to increase their size, will lead to a clipped child
being shown.

Also it doesnt seem to handle scrolling on its own. Even though
there is a scroll mode that can be specified, or else I am yet
to crack its proper usage as well as proper usage of clutter.