ListBox Scroll P3 - Motion Event
Trap motion event so that scrolling can occur in real time as the
mouse based press button and move it over the listbox before
releasing the same is occuring.

It also updates the pressPos so that the scroll amount between motion
events is equally to the actual movement of the mouse and doesnt go
on building up.

NOTE: Later have to add the logic for touch event also.