ListBox scroll allow upto 0.5 secs between events
User has upto 0.5 seconds between pressing of the button and moving
the mouse to trigger the scroll.

Similarly the user can pause for upto 0.5 secs in the middle of the
scroll.

If user waits for more than 0.5 seconds, then they will have to start
the scroll by releasing and pressing the mouse key again.

Also added a MS suffix to the threshold variable, to indicate that it
is in milli seconds.

NOTE: As of now I dont check which mouse button was pressed.