ListBox itemClick use Release not Press event
As the event handling starts from the inner most widget/actor to
the parents/outward. So If the inner most actor eats up the event
then it wont propagate to the parent.

Also Press event occurs before the Scroll event, so cant work on
Press event immidiately. While as Release occurs after Scroll (if
there is one) so deciding to treat the release as a itemclick
or not can be decided easily. If there was no (or almost no) scroll
then treat the button release as a item click, however if there was
sufficient scroll, then treat it as a ListBox scroll and not a item
click.

In case of ListBox with scrolling, we have events to handle at two
levels

1. the button/item click on the items in the listbox. (Inner)

2. the scrolling of the list box items. (External)

TODO: Will have to make the itemclick related inner button release
event handler return without stopping the event propogation. This
will allow the scroll logic in the listbox mouse handler terminate
the scroll event in a graceful way.