ListBox DONT scroll beyond the beginning.
Dont allow the listbox to be scrolled beyond/before its beginning

i.e in a Horizontal ListBox, when pulling it towards the right,
stop once the starting of the listbox is visible at the left visible
end of the listbox viewport.

i.e in a Vertical ListBox, when pulling it down, stop once the start
of the listbox is visible at the top visible end of the listbox
viewport.