Avoid checking Shift mask in key handler
Now we just check for Control Mask (ctrl key press) and Capital
A or Q for handling the keyboard shortcuts in the test app.

We no longer check to see if Shift Mask is present or not.ie if
shift key is pressed or not.

So with this

* if Control Key is pressed and

* if Capital Q or A is generated either by pressing Shift key and
  Q or A key and OR by having caps lock enabled and directly pressing
  Q or A key (i.e q or a key).

then the corresponding logic is triggered.