Clear Scroll BlurEffect after a timeout
Add a clutter threads timeout callback function which gets called
after a specified timeout interval with the GUI lock held, so that
it can operate on the gui.

This inturn is used to clear any blurEffect still active on the
specified ListBox.

This should ensure that even if the user releases the button during
a listbox scroll, outside the listbox, there is a oppurtunity to
clear the blur effect.

NOTE: The way the logics are setup, if the user is still in the middle
of the scroll when the timeout occurs, even thou the callback will
clear the blur effect, the scroll logic will set the blur, if required
as part of the ongoing scroll.

NOTE: Noticed that threads_add_idle and threads_add_timeout correspond
to the threads_add_idle|timeout_full in c.

NOTE: Using Clutter.some_func_name.get_arguments() will give the arguments
expected by that function which is directly under Clutter. Without this
making out what threads_add_... refers too would have become a trial and error.