Replaced -1 with -99 defined as IGNORE
Now I use IGNORE (defined as -99) as the special value used to
convey to helper routines to ignore setting up of the position
or size wrt the actor.

-1 is used by Clutter in some places like ClutterText to indicate
that any previous size set should cleared and the size which
corresponds to text contained in ClutterText should be used and so

So to avoid reusing -1 to mean something else, this change as been
done.