LBinLB-2DScroll:P2: Press Release propagate
Allow button press and release events to also propagate up the
hierarchy, as far as the listbox mouse handler is concerned.

So that if a listbox is embedded within another listbox, having
the other orientation, then one can try and achieve scrolling
in either direction, inturn mapping to the right listbox.