CB: UI Template file and Content metadata files
Program will use a UI Template file to create the UI. Initially
there will be only a single UI template file for the program which
inturn sets up a predefined set of listboxes.

Content metadata file will give details of the content the user
will be allowed to browse using the content browser.

This can be used to specify the list of main categories of the
content available for the user like Movies, TV shows, Audio Albums,
...

It can also be used to specify the contents belonging to each of
these category inturn organised into few groups.