CB: Create main category listbox from ui.template
Now the listbox which provides the list of categories is create at
runtime from the ui.template. This ensures that the previous mentioned
mouse click / motion related issues dont occur.

Have added the sample content metadata file which is used as the main
content metadata file for testing, into git.

TODO: Need to add a parent gui_id to the LISTBOX entry in ui template
so that one can use it to either create global listboxes or listboxes
which reside in other listboxes.

TODO: The button release after a scroll event shouldnt trigger
itemclick, this needs to be handled. Have to use existing motion/scroll
related data or a new flag to identify when one is exiting scrolling
and inturn based on that, avoid triggering the itemclick.