CB: Create ListBoxes and LBs within LB
As mentioned in the last commit, now for each listbox one can
specify a parent gui id. This can inturn be used to create
listboxes within other listbox.

Also if PID is NONE, then the listbox is added as a child to the
stage.

With this now the default content browser ui is fully created from
ui template.