CB: Try get another screen in App/Clutter
Trying to create a 2nd stage for a different window/screen for
the app.

I am able to create a new stage and show it, but if I try to set
a event / signal handler the app crashes after I swithc to the
other stage.

Maybe I have to create one more event loop and or thread for the
2nd stage or ... need to look at handling multiple stages parallely
from a single app, if possible, later.