CB:TargetType CMA, for Audio album
If target type is CMA, then the application switches the screen to
the audio screen and then inturn loads the content metadata file
specified following the CMA.

A set of sample files created to test the same.

Also the audio ui has been updated a bit wrt the positioning of
its elements.