CB: Slightly better Debug UI Template outline drawings
Now for listboxes which are children of other listboxes, they will
be drawn within their parent listbox.

TODO:

However it doesnt account for more than one child listbox with in
a parent listbox. This leads to wrong outline drawing if the child
listboxes are positioned 0,0 so that at runtime it will be auto
positioned one after the other automatically by clutter.

Equally it doesnt account about padding between the items in the listbox.