CB: Generate audio album related test images
Created test images for audio screen elements so that they
match the size of ui element they belong to. So that one can
be sure that some layout alignment or ... is not creating any
issue.

This is used to check the audio screen ui elements related bit
off position or so.

It seems like the songs and title are bit shifted to left rather
than being centered (as required and specified)

Need to look into this later.