CM: Stop before trying to play new content
The logic tries to stop any existing playing media, just in case.

NOTE: The CGST Player logic seems to work independent of this, but
just as a good control flow doing this.

LATER: Not sure if overwriting of a existing variable, is leading
to a clean clearing of it, before the new entity takes its place
or ClutterGST or GST taking care of it or ... Realised havent looked
at Python's object management flow in ages. Also coming back to
GST after few years now, so learning from scratch, in a way.