CB: Video Mp4 target link handled now
Added the video screen entry into gScreens.

Renamed the cm var within init_screen to meta, to avoid confusion
with the cm named assigned to my cluttermedia module during import.

As video playback requires most of the screen and is directly started
from the main screen, without going into a specific album related
screen as in audio, so first the screen is switched to video screen
and then the video is played.

Have added cm.video to define the ui contents for the video screen.

In ubuntu had to set LIBVA_DRIVER_NAME=i965 to get vaapi decode plugin
to support h264 codec. iHD for some reason didnt work. Fedora seems
to use i965 by default. This is on one of my devel laptop.

Modified a cat cm file to refer to the single mp4 test file which is
currently available.

Also reset the ui.video title element to relate to the 800 screen
size.