CM: Media Type, Stop play if returning from video
Now the cluttermedia maintains a variable which indicates if currently
audio file is being played or video is being played.

In turn CB will stop the video playback, if the user is coming back
from video playback into main screen. While if a audio file is playing,
it will let the same to continue to play, when returning to main screen.

TODO/MAYBE:
May need to give some time after triggering stop, before switching the
screen. Or better still trap some signal before switching the screen.

Need to check why audio within mp4 is not playing. Maybe the required
audio codec is currently not installed in my machine.

Currently as there is no other h264 gst plugin other than vaapi, so
vaapi is used for decode. However if other h264 decode plugins are
installed, not sure if vaapi will continue to be used. May add explicit
pipeline/decode graph setup with vaapi in it to ensure, vaapi is
always used.